formatted code & extracted session hook to middleware module
