+++ lib/index.js
@@ -8,9 +8,9 @@
 'use strict';
 
 var express = require('express'),
-  session   = require('express-session'),
-  io        = require('socket.io'),
-  Router    = require('./router.js');
+    session   = require('express-session'),
+    io        = require('socket.io'),
+    Router    = require('./router.js');
 
 module.exports.init = function (options) {
 
@@ -24,17 +24,20 @@ module.exports.init = function (options) {
   options.resave = options.resave === undefined ? true : false;
   options.saveUninitialized = options.saveUninitialized === undefined ? true : false;
 
-  _app = express(); //load an express instance
-  _app.express = express; //expose raw express object for middlewares
+  _app = express(); // Load an express instance
+  _app.express = express; // Expose raw express object for middlewares
 
-  if (!options.https) //load http or https nodejs server
+  // Load http or https nodejs server
+  if (!options.https) {
     _server = require('http').createServer(_app);
-  else
+  } else {
     _server = require('https').createServer(options.https, _app);
+  }
 
-  if (!options.disableSession)
-    _app.use(session(options)); //enable session support in express
-
+  if (!options.disableSession) {
+    _app.use(session(options)); // Enable session support in express
+  }
+  
   _app.io = io(_server);
   _app.rawServer = _server;
 
@@ -67,7 +70,9 @@ module.exports.init = function (options) {
   _app.io.route = function (namespace, eventName, fn) {
     _router.add(namespace, eventName, fn);
   };
-
+  
+  _app.io.Router = Router
+  
   return _app;
 
 };
