+++ lib/router.js
@@ -1,54 +1,21 @@
 'use strict';
 
 var Router = function (app, options) {
-
+  
   var that        = this;
   var isListening = false;
-
-  //expose public
-
+  var sessionMiddleware = require('./sessionMiddleware')(options);
+  
+  // Expose public
   this.add = add;
   this.addListeners = addListeners;
 
   this.routes = [];
   this.nsps   = ['/'];
 
-  function _sessionHook(socket, next) {
-
-    if (socket.request.headers.cookie && !options.disableSession) {
-      var _cookies = require('cookie').parse(socket.request.headers.cookie);
-      if (_cookies[options.name]) {
-
-        //decode the cookie using session secret
-        var _sid = require('cookie-parser').signedCookie(_cookies[options.name], options.secret);
-
-        options.store.get(_sid, function (err, session) {
-          if (err || !session) {
-            setInvalidSession(socket);
-            return next();
-          }
-
-          //Adding properties
-          session.save = function (fn) {
-            options.store.set(_sid, this, (fn || function () { }));
-          };
-
-          socket.session = session;
-          next();
-        });
-
-      } else {
-        setInvalidSession(socket); next();
-      }
-    } else {
-      setInvalidSession(socket); next();
-    }
-
-  }
-
   // Base session hook
-  // namespaces register their own within 'add'
-  app.io.use(_sessionHook);
+  // Namespaces register their own within 'add'
+  app.io.use(sessionMiddleware);
 
   /**
    * Register a route in current router.
@@ -80,7 +47,7 @@ var Router = function (app, options) {
 
       // For each namespace, we need to register the session middleware
       // This has to be done outside of addListeners as other middleware may rely on this and can be registered before addListeners is called
-      app.io.of(namespace).use(_sessionHook);
+      app.io.of(namespace).use(sessionMiddleware);
 
     }
   }
@@ -106,8 +73,7 @@ var Router = function (app, options) {
   }
 
 
-  /** PRIVATE METHODS **/
-
+  /** PRIVATE METHOD **/
   function listen(namespace, socket) {
 
     for (var i = 0; i < that.routes.length; i++) {
@@ -125,15 +91,5 @@ var Router = function (app, options) {
   }
 
 };
-
-/** PRIVATE **/
-
-function setInvalidSession(socket) {
-  socket.session = {};
-  socket.session.save = function (callback) {
-    callback = callback || function () { };
-    callback('Session is not initialized. Init it with a classic express request.');
-  };
-}
 
 module.exports = Router;
